moved IdentifierField descriptor
